Fix latent bugs and make ML models lazy-loaded
Phase 0 (correctness):
- Fix datetime.now() AttributeError in character/passage services
  (import datetime -> from datetime import datetime)
- Guard daily_stats streak null-deref on first-ever streak day;
  first-day streak length now 1, not 0
- difficulty column INTEGER -> FLOAT in migration (Postgres-safe)
- Add minimal runnable SRS helper self-check (tests/test_helpers.py)

Phase 5a (small-server memory):
- utils/models.py: @cache'd getters, nothing loads at import
- detect_language.py: lazy Whisper detection model
- spacy_model.py: lru_cache(maxsize=4)
- evaluator/feedback/text_gen/tts: model/tokenizer as lazy @property
- Remove dead deps uvicorn, python-multipart

Idle app RSS dropped from ~6-8GB to ~440MB (full app boots, /health 200).

+++ docs/migrations/2026_04_lesson_migration.sql
@@ -149,7 +149,7 @@ CREATE TABLE word (
         ELSE 'mastered'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- word-specific
     word TEXT NOT NULL,
     translation TEXT,
@@ -178,7 +178,7 @@ CREATE TABLE character (
         ELSE 'mastered'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- character-specific
     character TEXT NOT NULL,
     phonetic TEXT,
@@ -207,7 +207,7 @@ CREATE TABLE passage (
         ELSE 'mastered'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- passage-specific
     text TEXT NOT NULL,
     translation TEXT,
@@ -233,7 +233,7 @@ CREATE TABLE vocabulary (
         ELSE 'completed'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- relations
     lesson_id TEXT NOT NULL,
     word_id TEXT NOT NULL,
@@ -255,7 +255,7 @@ CREATE TABLE grammar (
         ELSE 'completed'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- grammar-specific
     title TEXT,
     explanation TEXT,
@@ -278,7 +278,7 @@ CREATE TABLE calligraphy (
         ELSE 'completed'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- relations
     lesson_id TEXT NOT NULL,
     character_id TEXT NOT NULL,
@@ -300,7 +300,7 @@ CREATE TABLE exercise (
         ELSE 'completed'
     END) VIRTUAL,
     score INTEGER NOT NULL DEFAULT 0,
-    difficulty INTEGER NOT NULL DEFAULT 0.5,
+    difficulty FLOAT NOT NULL DEFAULT 0.5,
     -- exercise-specific
     exercise_type TEXT NOT NULL,
     question TEXT NOT NULL,
